pyproject.toml
[tool.poetry]
name = "web-chat"
version = "0.1.0"
description = ""
authors = ["Pavel Kletukhin <paulfalcone@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
jinja2 = "^3.1.3"
uvicorn = "^0.27.1"
black = "^24.2.0"
aiofiles = "^23.2.1"
redis = "^5.0.3"
websockets = "^12.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
pydantic = "^2.6.4"
email-validator = "^2.1.1"
celery = {extras = ["rabbitmq"], version = "^5.3.6"}
flower = "^2.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
